+++ src/Phan/Plugin/Internal/ThrowAnalyzerPlugin.php
@@ -6,6 +6,7 @@ use ast;
 use ast\Node;
 use Phan\AST\ContextNode;
 use Phan\AST\UnionTypeVisitor;
+use Phan\CodeBase;
 use Phan\Config;
 use Phan\Exception\CodeBaseException;
 use Phan\Exception\IssueException;
@@ -13,16 +14,18 @@ use Phan\Exception\NodeException;
 use Phan\Issue;
 use Phan\Language\Context;
 use Phan\Language\Element\FunctionInterface;
+use Phan\Language\Element\Method;
 use Phan\Language\Type;
 use Phan\Language\UnionType;
 use Phan\PluginV3;
+use Phan\PluginV3\AnalyzeMethodCapability;
 use Phan\PluginV3\PluginAwarePostAnalysisVisitor;
 use Phan\PluginV3\PostAnalyzeNodeCapability;
 
 /**
  * Analyzes throw statements and compares them against the phpdoc (at)throws annotations
  */
-class ThrowAnalyzerPlugin extends PluginV3 implements PostAnalyzeNodeCapability
+class ThrowAnalyzerPlugin extends PluginV3 implements PostAnalyzeNodeCapability, AnalyzeMethodCapability
 {
     /**
      * This is invalidated every time this plugin is loaded (e.g. for tests)
@@ -37,6 +40,41 @@ class ThrowAnalyzerPlugin extends PluginV3 implements PostAnalyzeNodeCapability
             return ThrowRecursiveVisitor::class;
         }
         return ThrowVisitor::class;
+    }
+
+    /**
+     * Check for throw statements in __toString()
+     *
+     * @param CodeBase $code_base
+     * The code base in which the method exists
+     *
+     * @param Method $method
+     * A method being analyzed
+     *
+     * @override
+     */
+    public function analyzeMethod(
+        CodeBase $code_base,
+        Method $method
+    ) : void {
+        if (Config::get_closest_target_php_version_id() >= 70400) {
+            return;
+        }
+        if (strcasecmp($method->getName(), '__toString') !== 0) {
+            return;
+        }
+        $throws_union_type = $method->getThrowsUnionType();
+        if ($throws_union_type->isEmpty()) {
+            return;
+        }
+        Issue::maybeEmit(
+            $code_base,
+            $method->getContext(),
+            Issue::ThrowCommentInToString,
+            $method->getContext()->getLineNumberStart(),
+            $method->getRepresentationForIssue(),
+            $throws_union_type
+        );
     }
 }
 
@@ -63,14 +101,22 @@ class ThrowVisitor extends PluginAwarePostAnalysisVisitor
         $code_base = $this->code_base;
 
         $union_type = UnionTypeVisitor::unionTypeFromNode($code_base, $context, $node->children['expr']);
-        $union_type = $this->withoutCaughtUnionTypes($union_type);
+        $union_type = $this->withoutCaughtUnionTypes($union_type, true);
         if ($union_type->isEmpty()) {
             // Give up if we don't know
-            // TODO: Infer Throwable, if the original $union_type was empty
-            // and there are no try/catch blocks wrapping this.
             return;
         }
         $analyzed_function = $context->getFunctionLikeInScope($code_base);
+        if (Config::get_closest_target_php_version_id() < 70400) {
+            if ($analyzed_function instanceof Method && \strcasecmp('__toString', $analyzed_function->getName()) === 0) {
+                $this->emitIssue(
+                    Issue::ThrowStatementInToString,
+                    $node->lineno,
+                    $analyzed_function->getRepresentationForIssue(),
+                    (string)$union_type
+                );
+            }
+        }
 
         // TODO: This seems like it didn't work for A::c(A::d()) - See #1960 (InvalidArgumentException wasn't detected)
         foreach ($this->parent_node_list as $parent) {
@@ -93,11 +139,20 @@ class ThrowVisitor extends PluginAwarePostAnalysisVisitor
         $this->warnAboutPossiblyThrownType($node, $analyzed_function, $union_type);
     }
 
-    protected function withoutCaughtUnionTypes(UnionType $union_type) : UnionType
+    protected function withoutCaughtUnionTypes(UnionType $union_type, bool $is_raw_throw) : UnionType
     {
         if ($union_type->isEmpty()) {
-            // Give up if we don't know
-            return $union_type;
+            if (!$is_raw_throw) {
+                return $union_type;
+            }
+            // Infer Throwable, if the original $union_type was empty
+            // and there are no try/catch blocks wrapping this throw statement.
+            foreach ($this->parent_node_list as $parent) {
+                if ($parent->kind === ast\AST_TRY) {
+                    return $union_type;
+                }
+            }
+            return UnionType::fromFullyQualifiedRealString('\Throwable');
         }
 
         foreach ($this->parent_node_list as $parent) {
@@ -237,7 +292,7 @@ class ThrowRecursiveVisitor extends ThrowVisitor
                 $this->warnAboutPossiblyThrownType(
                     $node,
                     $analyzed_function,
-                    $this->withoutCaughtUnionTypes($invoked_function->getThrowsUnionType())
+                    $this->withoutCaughtUnionTypes($invoked_function->getThrowsUnionType(), false)
                 );
             }
         } catch (CodeBaseException $_) {
@@ -281,7 +336,7 @@ class ThrowRecursiveVisitor extends ThrowVisitor
         $this->warnAboutPossiblyThrownType(
             $node,
             $analyzed_function,
-            $this->withoutCaughtUnionTypes($invoked_method->getThrowsUnionType()),
+            $this->withoutCaughtUnionTypes($invoked_method->getThrowsUnionType(), false),
             $invoked_method
         );
     }
@@ -321,7 +376,7 @@ class ThrowRecursiveVisitor extends ThrowVisitor
         $this->warnAboutPossiblyThrownType(
             $node,
             $analyzed_function,
-            $this->withoutCaughtUnionTypes($invoked_method->getThrowsUnionType()),
+            $this->withoutCaughtUnionTypes($invoked_method->getThrowsUnionType(), false),
             $invoked_method
         );
     }